Add debug mode to tempdir().
 * This helps in debugging temporary objects created during compile
 time.
 * Only consult DEBUG_MODE once at TempDirectory create time.

+++ python/tvm/contrib/util.py
@@ -16,8 +16,10 @@
 # under the License.
 """Common system utilities"""
 import atexit
+import datetime
 import os
 import tempfile
+import threading
 import shutil
 try:
     import fcntl
@@ -29,6 +31,31 @@ class TempDirectory(object):
 
     Automatically removes the directory when it went out of scope.
     """
+
+    # When True, all TempDirectory are *NOT* deleted and instead live inside a predicable directory tree.
+    DEBUG_MODE = False
+
+    # In debug mode, each tempdir is named after the sequence
+    _NUM_TEMPDIR_CREATED = 0
+    _NUM_TEMPDIR_CREATED_LOCK = threading.Lock()
+    @classmethod
+    def _increment_num_tempdir_created(cls):
+        with cls._NUM_TEMPDIR_CREATED_LOCK:
+            to_return = cls._NUM_TEMPDIR_CREATED
+            cls._NUM_TEMPDIR_CREATED += 1
+
+        return to_return
+
+    _DEBUG_PARENT_DIR = None
+    @classmethod
+    def _get_debug_parent_dir(cls):
+        if cls._DEBUG_PARENT_DIR is None:
+            all_parents = f'{os.environ["TMPDIR"]}/tvm-debug-mode-tempdirs'
+            if not os.path.isdir(all_parents):
+                os.makedirs(all_parents)
+            cls._DEBUG_PARENT_DIR = tempfile.mkdtemp(
+                prefix=datetime.datetime.now().strftime('%Y-%m-%dT%H-%M-%S___'), dir=all_parents)
+        return cls._DEBUG_PARENT_DIR
 
     TEMPDIRS = set()
     @classmethod
@@ -43,19 +70,27 @@ class TempDirectory(object):
         cls.TEMPDIRS = None
 
     def __init__(self, custom_path=None):
+        self._created_with_debug_mode = self.DEBUG_MODE
         if custom_path:
             os.mkdir(custom_path)
             self.temp_dir = custom_path
         else:
-            self.temp_dir = tempfile.mkdtemp()
+            if self._created_with_debug_mode:
+                parent_dir = self._get_debug_parent_dir()
+                self.temp_dir = f'{parent_dir}/{self._increment_num_tempdir_created():05d}'
+                os.mkdir(self.temp_dir)
+            else:
+                self.temp_dir = tempfile.mkdtemp()
 
-        self.TEMPDIRS.add(self.temp_dir)
+        if not self._created_with_debug_mode:
+            self.TEMPDIRS.add(self.temp_dir)
 
     def remove(self):
         """Remote the tmp dir"""
         if self.temp_dir:
-            shutil.rmtree(self.temp_dir, ignore_errors=True)
-            self.TEMPDIRS.remove(self.temp_dir)
+            if not self._created_with_debug_mode:
+                shutil.rmtree(self.temp_dir, ignore_errors=True)
+                self.TEMPDIRS.remove(self.temp_dir)
             self.temp_dir = None
 
     def __del__(self):
